- point entry - is dir condition
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@14405 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tine20/Tinebase/Helper.php
@@ -117,11 +117,11 @@ function removeDir($_dir)
 {
     $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator($_dir), RecursiveIteratorIterator::CHILD_FIRST);
     foreach ($iterator as $fullFilename => $cur) {
-        if (is_dir($fullFilename)) {
+        if (is_dir($fullFilename) && substr($fullFilename, -1) !== '.') {
             rmdir($fullFilename);
-        } else {
+        } elseif (is_file($fullFilename)) {
             unlink($fullFilename);
         }
     }
-    rmdir($_dir);
+    if (is_dir($_dir)) rmdir($_dir);
 }
